applets: Fix the manager proxy
Also remove the needless checks to see if the sender is actually the
currently active applet. We don't really need to worry about this for
all of the methods available, save one: Present is the only one that
actually needs to validate the applet requesting the presentation is
actually the one on top. Otherwise chaos will ensue.